Add 'Start Step' button for available task
